+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+* Added SpWML widget.
+* Added StackElement.
+* Fixed bug at replaceID function in SpWMLBuilder.
+
 ## 0.0.2
 
 * Update README.